Add support for sDctEL1252
Do a test to check how good the synchronization between the
timing-sytem -> PTP server -> EL6688 -> Distributed Clock
is.

We have 2 signals from the timing system:
- the PTP "signal", connected to the EL6688
- a 1 Hz pulse, feed into input 1 of an EL1252-0050

With each raising edge of the 1 Hz pulse the time is latched inside the
terminal, and then nothing happens within the next second.
The IOC reads the time into an EPICS PV.
The time is seconds.nanoseconds, and since we have a 1 Hz pulse,
the nanosecond part of the time should stay "constant".
We need to find out, how constant it really is.

Define a template and a proto file to read this timing as an ASCII string,
which is good enough for the moment.