simulator/hw_motor.c: Use MotorPosReported in log file
When simulating a motor, it may happen that the motor
is not exact on the same position, because of the encoder resolution:
The motor stops at 2.010 mm instead of 2.000
Note: this is dependent on the settings of the simulator, it needs
the "resolution" to be configured.

Anyway, use
"MotorPosReported" instead of "MotorPosNow" when writing to the log files.