Support for writing HomProc and HomPos
Add 2 records to all testing of homing procedures:
"-HomProc" and "-HomPos"
Both are localed in the debug template.
Note that writing these paramaters must still be allowed from
the motion controller.
Having these new records does not guarantee that we can mess with the
homing on all devices - this is still reserved for testing and/or
test crates.

Changes to be committed:
    modified:   ethercatmcApp/Db/ethercatmcdebug.template
    modified:   ethercatmcApp/src/ethercatmcIndexerAxis.cpp